Function declaration cleanup, eslint cleanup
