+++ .eslintrc.json
@@ -12,7 +12,8 @@
         "sourceType": "module"
     },
     "plugins": [
-        "react"
+        "react",
+        "react-hooks"
     ],
     "rules": {
         "indent": [
@@ -31,6 +32,9 @@
             "error",
             "always"
         ],
+        "no-unused-vars": "warn",
+        "no-unused-expressions": "warn",
+        "indent-legacy": "warn",
         "react-hooks/rules-of-hooks": "error", // Checks rules of Hooks
         "react-hooks/exhaustive-deps": "warn" // Checks effect dependencies
     }
